+++ src/types/Submission.d.ts
@@ -4,10 +4,10 @@ export class Submission {
   code: string;
   timestamp: number;
   statusCode: number;
-  runtimeDisplay: string;
-  runtimePercentile: number;
-  memoryDisplay: string;
-  memoryPercentile: number;
+  runtimeDisplay?: string | null;
+  runtimePercentile?: number | null;
+  memoryDisplay?: string | null;
+  memoryPercentile?: number | null;
   notes?: string;
   lang: {
     verboseName: string;